requirements.txt
pyyaml>=6.0
jsonschema>=4.22.0
